Add flag to manage architecture code dependant.
