+++ src/data_provider/src/hardware/hardwareWrapperImplMac.h
@@ -21,9 +21,8 @@
 #include "osPrimitivesInterfaceMac.h"
 #include "utilsWrapperMac.hpp"
 
-#if (MAC_OS_X_VERSION_MAX_ALLOWED < 120000) // Before macOS 12 Monterey
-#define kIOMainPortDefault kIOMasterPortDefault
-#endif
+
+double getMhz(IOsPrimitivesMac* osPrimitives = nullptr);
 
 template <class TOsPrimitivesMac>
 class OSHardwareWrapperMac final : public IOSHardwareWrapper, public TOsPrimitivesMac
@@ -101,160 +100,9 @@ class OSHardwareWrapperMac final : public IOSHardwareWrapper, public TOsPrimitiv
             return cores;
         }
 
-        uint64_t getCpuHzAarch64() const
+        double cpuMhz()
         {
-            uint64_t cpuHz = 0;
-
-            auto matching = this->IOServiceMatching("AppleARMIODevice");
-
-            if (matching == nullptr)
-            {
-                throw std::system_error
-                {
-                    0,
-                    std::system_category(),
-                    "Error on library function call IOServiceMatching."
-                };
-            }
-
-            io_iterator_t device_it = 0;
-            auto kr = this->IOServiceGetMatchingServices(kIOMainPortDefault, matching, &device_it);
-
-            if (kr != KERN_SUCCESS)
-            {
-                throw std::system_error
-                {
-                    kr,
-                    std::system_category(),
-                    "Error on library function call IOServiceGetMatchingServices."
-                };
-            }
-
-            DEFER([this, device_it]()
-            {
-                this->IOObjectRelease(device_it);
-            });
-
-            io_object_t device = 0;
-
-            while ((device = this->IOIteratorNext(device_it)))
-            {
-                DEFER([this, device]()
-                {
-                    this->IOObjectRelease(device);
-                });
-
-                io_name_t buf;
-                kr = this->IORegistryEntryGetName(device, buf);
-
-                if (kr != KERN_SUCCESS)
-                {
-                    continue;
-                }
-
-                std::string name(buf);
-
-                if (name.compare("pmgr"))
-                {
-                    continue;
-                }
-
-                CFMutableDictionaryRef properties;
-                kr = this->IORegistryEntryCreateCFProperties(device, &properties, kCFAllocatorDefault, kNilOptions);
-
-                if (kr != KERN_SUCCESS)
-                {
-                    throw std::system_error
-                    {
-                        kr,
-                        std::system_category(),
-                        "Error on library function call IORegistryEntryCreateCFProperties."
-                    };
-                }
-
-                DEFER([this, properties]()
-                {
-                    this->CFRelease(properties);
-                });
-
-                // voltage-states5-sram contains the performance cores available frequencies
-                CFStringRef cfkey = this->CFStringCreateWithCString(kCFAllocatorDefault, "voltage-states5-sram", kCFStringEncodingUTF8);
-                DEFER([this, cfkey]()
-                {
-                    this->CFRelease(cfkey);
-                });
-
-                auto p_cores_freq_property = static_cast<CFDataRef>(this->CFDictionaryGetValue(properties, cfkey));
-
-                if (p_cores_freq_property == nullptr)
-                {
-                    throw std::system_error
-                    {
-                        0,
-                        std::system_category(),
-                        "Error on library function call CFDictionaryGetValue."
-                    };
-                }
-
-                auto p_cores_freq_type = this->CFGetTypeID(p_cores_freq_property);
-
-                if (p_cores_freq_type != this->CFDataGetTypeID())
-                {
-                    throw std::system_error
-                    {
-                        0,
-                        std::system_category(),
-                        "CF type id of p_cores_freq_property is not Data type id."
-                    };
-                }
-
-                size_t length = this->CFDataGetLength(p_cores_freq_property);
-
-                // The frequencies are in hz, saved in an array as little endian 4 byte integers
-                for (size_t i = 0; i < length - 3; i += sizeof(uint32_t))
-                {
-                    uint32_t cur_freq = 0;
-                    this->CFDataGetBytes(p_cores_freq_property, this->CFRangeMake(i, sizeof(uint32_t)), reinterpret_cast<UInt8*>(&cur_freq));
-
-                    if (cpuHz < cur_freq)
-                    {
-                        cpuHz = cur_freq;
-                    }
-                }
-            }
-
-            return cpuHz;
-        }
-
-        double cpuMhz() const
-        {
-            constexpr auto MHz{1000000};
-            uint64_t cpuHz{0};
-            size_t len{sizeof(cpuHz)};
-            int ret{this->sysctlbyname("hw.cpufrequency", &cpuHz, &len, nullptr, 0)};
-
-            if (ret)
-            {
-                try
-                {
-                    cpuHz = getCpuHzAarch64();
-                }
-                catch (std::system_error& e)
-                {
-                    throw e;
-                }
-                catch (...)
-                {
-                    throw std::system_error
-                    {
-                        ret,
-                        std::system_category(),
-                        "Error reading cpu frequency."
-                    };
-                }
-            }
-
-            return static_cast<double>(cpuHz) / MHz;
+            return getMhz(static_cast<IOsPrimitivesMac*>(this));
         }
 
         uint64_t ramTotal() const
@@ -314,4 +162,6 @@ class OSHardwareWrapperMac final : public IOSHardwareWrapper, public TOsPrimitiv
             return 100 - (100 * ramFree() / ramTotal());
         }
 };
+
+
 #endif // _HARDWARE_WRAPPER_IMPL_MAC_H
